move common methods to SpecWorker and tidy imports
